chore: consolidate db migration scripts into shared scripts/db directory
