+++ connectors/package.json
@@ -17,15 +17,17 @@
     "test:ci": "vitest --reporter=junit --outputFile=junit.xml --watch=false",
     "tsgo": "tsgo",
     "cli": "npx tsx src/admin/cli.ts",
-    "migration:generate:pre-deploy": "./scripts/generate-migration.sh pre-deploy",
-    "migration:generate:post-deploy": "./scripts/generate-migration.sh post-deploy",
-    "migration:apply:pre-deploy": "node scripts/run-migrate.cjs --command pre-deploy --execute",
-    "migration:apply:post-deploy": "node scripts/run-migrate.cjs --command post-deploy --execute",
+    "initdb": "./admin/init_db.sh",
+    "create-db-migration": "../scripts/db/create_db_migration_file.sh",
+    "migration:generate:pre-deploy": "MIGRATION_DB_URI_VAR=CONNECTORS_DATABASE_URI MIGRATION_SHADOW_PREFIX=dust_connectors_shadow_to MIGRATION_ADMIN_DB_PATH=src/admin/db.ts ../scripts/db/generate-migration.sh pre-deploy",
+    "migration:generate:post-deploy": "MIGRATION_DB_URI_VAR=CONNECTORS_DATABASE_URI MIGRATION_SHADOW_PREFIX=dust_connectors_shadow_to MIGRATION_ADMIN_DB_PATH=src/admin/db.ts ../scripts/db/generate-migration.sh post-deploy",
+    "migration:apply:pre-deploy": "node ../scripts/db/run-migrate.cjs --command pre-deploy --execute",
+    "migration:apply:post-deploy": "node ../scripts/db/run-migrate.cjs --command post-deploy --execute",
     "migration:apply": "npm run migration:apply:pre-deploy && npm run migration:apply:post-deploy",
-    "migration:status": "node scripts/run-migrate.cjs --command status --execute",
-    "migration:check": "node scripts/run-migrate.cjs --command check --execute",
-    "migration:check:pre-deploy": "node scripts/run-migrate.cjs --command check-pre-deploy --execute",
-    "migration:check:post-deploy": "node scripts/run-migrate.cjs --command check-post-deploy --execute",
+    "migration:status": "node ../scripts/db/run-migrate.cjs --command status --execute",
+    "migration:check": "node ../scripts/db/run-migrate.cjs --command check --execute",
+    "migration:check:pre-deploy": "node ../scripts/db/run-migrate.cjs --command check-pre-deploy --execute",
+    "migration:check:post-deploy": "node ../scripts/db/run-migrate.cjs --command check-post-deploy --execute",
     "teams:create-app": "tsx ./scripts/create-teams-app.ts",
     "discord:register-commands": "tsx ./scripts/register-discord-commands.ts",
     "knip": "knip"
