+++ dockerfiles/connectors.Dockerfile
@@ -24,6 +24,9 @@ COPY /connectors/ .
 RUN find . -name "*.test.ts" -delete
 RUN find . -name "*.test.tsx" -delete
 
+# Copy shared migration tooling (scripts/migrate.ts imports ../../scripts/db/migration-runner)
+COPY /scripts/db /app/scripts/db
+
 # Build temporal workers
 RUN CONNECTORS_DATABASE_URI="postgres://fake:fake@localhost:5432/fake" npm run build:temporal-bundles
 # Build all components (server, worker, cli) with esbuild
