+++ front/package.json
@@ -23,16 +23,16 @@
     "test:ci": "vitest --reporter=junit --watch=false --test-timeout=30000",
     "coverage": "vitest --coverage",
     "initdb": "./admin/init_db.sh",
-    "create-db-migration": "./create_db_migration_file.sh",
-    "migration:generate:pre-deploy": "./scripts/generate-migration.sh pre-deploy",
-    "migration:generate:post-deploy": "./scripts/generate-migration.sh post-deploy",
-    "migration:apply:pre-deploy": "node scripts/run-migrate.cjs --command pre-deploy --execute",
-    "migration:apply:post-deploy": "node scripts/run-migrate.cjs --command post-deploy --execute",
+    "create-db-migration": "../scripts/db/create_db_migration_file.sh",
+    "migration:generate:pre-deploy": "MIGRATION_DB_URI_VAR=FRONT_DATABASE_URI MIGRATION_SHADOW_PREFIX=dust_front_shadow_to MIGRATION_ADMIN_DB_PATH=admin/db.ts ../scripts/db/generate-migration.sh pre-deploy",
+    "migration:generate:post-deploy": "MIGRATION_DB_URI_VAR=FRONT_DATABASE_URI MIGRATION_SHADOW_PREFIX=dust_front_shadow_to MIGRATION_ADMIN_DB_PATH=admin/db.ts ../scripts/db/generate-migration.sh post-deploy",
+    "migration:apply:pre-deploy": "node ../scripts/db/run-migrate.cjs --command pre-deploy --execute",
+    "migration:apply:post-deploy": "node ../scripts/db/run-migrate.cjs --command post-deploy --execute",
     "migration:apply": "npm run migration:apply:pre-deploy && npm run migration:apply:post-deploy",
-    "migration:status": "node scripts/run-migrate.cjs --command status --execute",
-    "migration:check": "node scripts/run-migrate.cjs --command check --execute",
-    "migration:check:pre-deploy": "node scripts/run-migrate.cjs --command check-pre-deploy --execute",
-    "migration:check:post-deploy": "node scripts/run-migrate.cjs --command check-post-deploy --execute",
+    "migration:status": "node ../scripts/db/run-migrate.cjs --command status --execute",
+    "migration:check": "node ../scripts/db/run-migrate.cjs --command check --execute",
+    "migration:check:pre-deploy": "node ../scripts/db/run-migrate.cjs --command check-pre-deploy --execute",
+    "migration:check:post-deploy": "node ../scripts/db/run-migrate.cjs --command check-post-deploy --execute",
     "debug:profiler": "node ./scripts/debug/run_profiler.ts",
     "generate:hubspot-forms": "tsx scripts/generate-hubspot-forms.ts"
   },
